Approach 7: Use bytes instead of strings(Time: 58s)
The read_lines and read_line method stored the line as a String. And
since String in rust needs to be UTF8, the previous approach validated
this on each line.

We replace strings with bytes and only validate the string while
printing. Playing with bytes is also much cheaper because you are dealing
with raw data and no processing is required before you can access it.